+++ assets/css/main.css
@@ -18,6 +18,10 @@
 @import "../../common.blocks/button/button.css";
 @import "../../common.blocks/button/_focused/button_focused.css";
 @import "../../common.blocks/button-link/button-link.css";
+@import "../../common.blocks/button-link/button-link__back-top/button-link__back-top.css";
+@import "../../common.blocks/button-link/button-link_focused/button-link_focused.css";
+@import "../../common.blocks/button-link/button-link__back-top/button-link__back-top_focused/button-link__back-top_focused.css";
+
 
 @import "../../desktop.blocks/button/button.css";
 
@@ -25,7 +29,6 @@
 @import "../../touch.blocks/button/button.css";
 @import "../../touch.blocks/button/_focused/button_focused.css";
 @import "../../touch.blocks/button/_hide/buttun_hide.css";
-
 /* hr {
   border: none; 
   background-color: black;
