+++ assets/js/main.js
@@ -1,6 +1,5 @@
-import { backToTop } from "../../common.blocks/button-link/back-to-top.js";
-import { copyright } from '../../common.blocks/copyright/copyright.js';
+import { copyright } from "../../common.blocks/copyright/copyright.js";
+import { backToTop } from "../../common.blocks/button-link/button-link__back-top/button-link__back-top.js";
 
-backToTop();
 copyright();
-
+backToTop();
