+++ common.blocks/button-link/button-link__back-top/button-link__back-top.css
@@ -1,0 +1,15 @@
+.rounded-link-button {
+  position: fixed;
+  right: 15px;
+  bottom: 15px;
+  z-index: 999;
+}
+
+.arrow-up {
+  background-image:  url("/assets/img/arrow-up-white.png");
+  background-size: contain;
+  background-repeat: no-repeat;
+  width: 20px;
+  height: 20px;
+  background-position: center;
+}
