+++ common.blocks/button-link/button-link__back-top/button-link__back-top_focused/button-link__back-top_focused.css
@@ -1,0 +1,3 @@
+.arrow-up:hover {
+    background-image: url("/assets/img/arrow-up-red.png");
+  }
