+++ common.blocks/button-link/button-link_focused/button-link_focused.css
@@ -1,10 +1,5 @@
-.btn-active:hover button {
+.rounded-link-button:hover {
   color: var(--nine-global-color-primary);
   background: var(--nine-global-color-background-light);
-  border: 2px solid var(--nine-global-color-primary);
-  min-width: 100%;
-}
-
-#footer .btn-active:hover button {
   border: 2px solid var(--nine-global-color-primary);
 }
