+++ index.html
@@ -619,7 +619,7 @@
     <!-- Back to top button -->
     <div id="back-to-top">
     <a  href="#" class="rounded-link-button d-flex align-items-center justify-content-center"><i
-            class="bi bi-arrow-up-short"></i></a>
+            class="arrow-up"></i></a>
         </div>
     <!-- End Back to top button -->
